# konto-verwaltung-ui/src/app/api-services/models/month-total.ts
/* tslint:disable */
/* eslint-disable */
export interface MonthTotal {
  month?: 'JANUARY' | 'FEBRUARY' | 'MARCH' | 'APRIL' | 'MAY' | 'JUNE' | 'JULY' | 'AUGUST' | 'SEPTEMBER' | 'OCTOBER' | 'NOVEMBER' | 'DECEMBER';
  total?: number;
}