# konto-verwaltung-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Minh Chi Diep",
    "description": "OpenApi documentation for account management",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Minh Chi Diep",
      "url": "https://minhchi1709.github.io/mywebsite/index.html",
      "email": "diepminhchi1617@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088",
      "description": "Local ENV"
    },
    {
      "url": "https://aliboucoding.com/course",
      "description": "PROD ENV"
    }
  ],
  "paths": {
    "/app/AccountManagement/api/v1/transactions/transactions/{id}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "getTransactionById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Transaction"
        ],
        "operationId": "updateTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Transaction"
        ],
        "operationId": "deleteTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/transactions": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "getAllTransactions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Transaction"
        ],
        "operationId": "createTransaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/transactions/years": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "getAllYears",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/transactions/years/{year}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "getAllTransactionsOfYear",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/transactions/years/{year}/months": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "getAllMonths",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "enum": [
                      "JANUARY",
                      "FEBRUARY",
                      "MARCH",
                      "APRIL",
                      "MAY",
                      "JUNE",
                      "JULY",
                      "AUGUST",
                      "SEPTEMBER",
                      "OCTOBER",
                      "NOVEMBER",
                      "DECEMBER"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/transactions/years/{year}/months/{month}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "getAllTransactionsOfMonth",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/transactions/years/{year}/month-total": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "getAllMonthTotals",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonthTotal"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/currency": {
      "get": {
        "tags": [
          "Currency"
        ],
        "operationId": "getAllRates",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CurrencyRate"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/currency/vib": {
      "get": {
        "tags": [
          "Currency"
        ],
        "operationId": "getAllVIBRates",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CurrencyRateVIB"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/currency/today": {
      "get": {
        "tags": [
          "Currency"
        ],
        "operationId": "getTodayCurrency",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyRate"
                }
              }
            }
          }
        }
      }
    },
    "/app/AccountManagement/api/v1/currency/today/vib": {
      "get": {
        "tags": [
          "Currency"
        ],
        "operationId": "getTodayVIBCurrency",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyRateVIB"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TransactionRequest": {
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "format": "float"
          },
          "date": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "type": "number",
            "format": "float"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "MonthTotal": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string",
            "enum": [
              "JANUARY",
              "FEBRUARY",
              "MARCH",
              "APRIL",
              "MAY",
              "JUNE",
              "JULY",
              "AUGUST",
              "SEPTEMBER",
              "OCTOBER",
              "NOVEMBER",
              "DECEMBER"
            ]
          },
          "total": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "CurrencyRate": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "rate": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "CurrencyRateVIB": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "rate": {
            "type": "number",
            "format": "float"
          }
        }
      }
    }
  }
}